+++ packages/babel/src/index.ts
@@ -7,7 +7,7 @@ import { buildPlugins, buildProductionPlugins } from './plugins.js';
 import { buildPresets } from './presets.js';
 import { applyUserConfig } from './user-config.js';
 
-export type { CreateBabelPresetsOptions } from './types.js';
+export type { BabelMergeContext, BabelMergeFunction, CreateBabelPresetsOptions, Framework, Modules } from './types.js';
 
 export const createBabelPresets = ({
   framework = 'none',